requirements.txt
Flask==3.0.2
SQLAlchemy>=2.0.30
Flask-Limiter==3.5.0
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.1
httpx==0.26.0
yt-dlp==2025.1.15
apify-client>=1.6.0
openai-whisper
sumy==0.11.0
nltk==3.8.1
